+++ odoo/local-src/specific_sale/tests/test_sale_order_line_qty_unavailable.py
@@ -3,6 +3,9 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 from odoo.tests.common import TransactionCase, at_install, post_install
+from datetime import timedelta
+from odoo.fields import Datetime
+
 
 import logging
 
@@ -71,6 +74,7 @@ class TestSaleOrderLineQtyUnavailable(TransactionCase):
         # Create the first sale order with 10 as ordered quantity
         self.sale_1 = self.env['sale.order'].create({
             'partner_id': self.partner.id,
+            'date_order': Datetime.now(),
             'order_line': [
                 (0, 0, {
                     'name': self.p1.name,
@@ -119,6 +123,9 @@ class TestSaleOrderLineQtyUnavailable(TransactionCase):
         # Create the second sale order with 5 as ordered quantity
         self.sale_2 = self.env['sale.order'].create({
             'partner_id': self.partner.id,
+            'date_order': Datetime.to_string(
+                Datetime.from_string(Datetime.now()) + timedelta(hours=1)
+            ),
             'order_line': [
                 (0, 0, {
                     'name': self.p1.name,
@@ -188,59 +195,11 @@ class TestSaleOrderLineQtyUnavailable(TransactionCase):
         )
         # After the stock increase (qty = 2),
         # the unavailable quantity on first order is already 10
-        # FIXME randomly get 10 instead of 8
-        _logger.info("======== Start debug info for unit test =========")
-        _logger.info("count lines: %s", len(self.sale_1.order_line[0]))
-        line = self.sale_1.order_line[0]
-        _logger.info("product_id: %s", line.product_id)
-        _logger.info("product_uom_qty: %s",
-                     line.product_uom_qty)
-        _logger.info("state: %s", line.state)
-        _logger.info("id: %s", line.id)
-        _logger.info("product.immediately_usable_qty: %s",
-                     line.product_id.immediately_usable_qty)
-        order_line_stock_move = self.env['stock.move'].search([
-            ('procurement_id.sale_line_id', '=', line.id),
-            ('state', 'not in', ['draft', 'cancel', 'done'])
-        ])
-        _logger.info("order_line_stock_move: %s",
-                     order_line_stock_move)
-        stock_move_date_expected = (
-            order_line_stock_move.date_expected
-        )
-        _logger.info("stock_move_date_expected: %s",
-                     stock_move_date_expected)
-        next_stock_moves = self.env['stock.move'].search([
-            ('product_id', '=', line.product_id.id),
-            ('procurement_id.sale_line_id', '!=', line.id),
-            ('state', 'not in', ['draft', 'cancel', 'done']),
-            '|',
-            ('priority', '<',  order_line_stock_move.priority),
-            '&',
-            ('priority', '=', order_line_stock_move.priority),
-            ('date_expected', '>', stock_move_date_expected),
-        ])
-        _logger.info("next_stock_moves: %s",
-                     next_stock_moves)
-        next_quantities = sum(
-            move.product_uom_qty for move in next_stock_moves
-        )
-        for move in next_stock_moves:
-            _logger.info("move %s qty: %s",
-                         move.id, move.product_uom_qty)
-
-        good_immediately_usable_qty = (
-            line.product_id.immediately_usable_qty + next_quantities
-        )
-        _logger.info("good_immediately_usable_qty: %s",
-                     good_immediately_usable_qty)
-        _logger.info("======== End debug info for unit test =========")
         self.assertEqual(
             self.sale_1.order_line[0].product_qty_unavailable,
             10
         )
         # the current unavailable quantity on first order is now 8
-
         self.assertEqual(
             self.sale_1.order_line[0].current_product_qty_unavailable,
             8